revert: moving map to modules folder
